test -  move shared fixture in common to conftest
