+++ test/common/unit/conftest.py
@@ -3,6 +3,7 @@ import random
 from shutil import rmtree
 
 from common.db_manager import DBManager
+from common.constants import LOCATOR_LEN_BYTES
 
 
 @pytest.fixture(scope="session", autouse=True)
@@ -19,6 +20,23 @@ def db_manager():
 
     manager.db.close()
     rmtree("test_db")
+
+
+@pytest.fixture
+def appointment_data():
+    locator = get_random_value_hex(LOCATOR_LEN_BYTES)
+    start_time = 100
+    end_time = 120
+    to_self_delay = 20
+    encrypted_blob_data = get_random_value_hex(100)
+
+    return {
+        "locator": locator,
+        "start_time": start_time,
+        "end_time": end_time,
+        "to_self_delay": to_self_delay,
+        "encrypted_blob": encrypted_blob_data,
+    }
 
 
 def get_random_value_hex(nbytes):
