+++ test/common/unit/test_appointment.py
@@ -1,33 +1,8 @@
 import struct
 import pytest
 import binascii
-import pyzbase32
-from pytest import fixture
-from coincurve import PrivateKey
 
 from common.appointment import Appointment
-from common.cryptographer import Cryptographer
-from common.constants import LOCATOR_LEN_BYTES
-
-from test.common.unit.conftest import get_random_value_hex
-
-
-# Not much to test here, adding it for completeness
-@fixture
-def appointment_data():
-    locator = get_random_value_hex(LOCATOR_LEN_BYTES)
-    start_time = 100
-    end_time = 120
-    to_self_delay = 20
-    encrypted_blob_data = get_random_value_hex(100)
-
-    return {
-        "locator": locator,
-        "start_time": start_time,
-        "end_time": end_time,
-        "to_self_delay": to_self_delay,
-        "encrypted_blob": encrypted_blob_data,
-    }
 
 
 def test_init_appointment(appointment_data):
